use ignore keyboard trick on floating panel, too
