+++ Examples/Maps-SwiftUI/Maps/HostingCell.swift
@@ -47,48 +47,3 @@ public final class HostingCell<Content: View>: UITableViewCell {
         }
     }
 }
-
-/// This extension fixes a safe area issue happening with `HostingCell` when a
-/// keyboard is shown.
-///
-/// The bug is present only in iOS 14.0-14.2 and only if a keyboard is shown.
-///
-/// Credits to https://steipete.me/posts/disabling-keyboard-avoidance-in-swiftui-uihostingcontroller/
-@available(iOS, introduced: 13, deprecated: 14.2, message: "No longer necessary after iOS 14.2.")
-extension UIHostingController {
-    public convenience init(rootView: Content, ignoresKeyboard: Bool) {
-        self.init(rootView: rootView)
-
-        if ignoresKeyboard {
-            guard let viewClass = object_getClass(view) else { return }
-
-            let viewSubclassName = String(
-                cString: class_getName(viewClass)
-            ).appending("_IgnoresKeyboard")
-
-            if let viewSubclass = NSClassFromString(viewSubclassName) {
-                object_setClass(view, viewSubclass)
-            } else {
-                guard
-                    let viewClassNameUtf8 = (viewSubclassName as NSString).utf8String,
-                    let viewSubclass = objc_allocateClassPair(viewClass, viewClassNameUtf8, 0)
-                else { return }
-
-                if let method = class_getInstanceMethod(
-                    viewClass,
-                    NSSelectorFromString("keyboardWillShowWithNotification:")
-                ) {
-                    let keyboardWillShow: @convention(block) (AnyObject, AnyObject) -> Void = { _, _ in }
-                    class_addMethod(
-                        viewSubclass,
-                        NSSelectorFromString("keyboardWillShowWithNotification:"),
-                        imp_implementationWithBlock(keyboardWillShow),
-                        method_getTypeEncoding(method)
-                    )
-                }
-                objc_registerClassPair(viewSubclass)
-                object_setClass(view, viewSubclass)
-            }
-        }
-    }
-}
